+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to the "vscode-camelk" extension will be documented in this 
 
 ## 0.0.34
 
+- Update default runtime version to v1.12.0
+
 ## 0.0.33
 
 - Update default runtime version to v1.11.0